per #344, move more functions out of met_util into a more appropriate util file
