+++ internal/tests/pytests/util/met_util/test_met_util.py
@@ -212,19 +212,6 @@ def test_get_lead_sequence_init_min_10(metplus_config):
 
 
 @pytest.mark.parametrize(
-    'value, expected_result', [
-        (3.3, 3.5),
-        (3.1, 3.0),
-        (-3.2, -3.0),
-        (-3.8, -4.0),
-    ]
-)
-@pytest.mark.util
-def test_round_0p5(value, expected_result):
-    assert util.round_0p5(value) == expected_result
-
-
-@pytest.mark.parametrize(
     'int_string, expected_result', [
         ('4', [4]),
         ('4-12', [4, 5, 6, 7, 8, 9, 10, 11, 12]),
@@ -259,101 +246,6 @@ def test_subset_list(subset_definition, expected_result):
     full_list = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j']
     result = util.subset_list(full_list, subset_definition)
     assert result == expected_result
-
-
-@pytest.mark.parametrize(
-    'filename, expected_result', [
-        # file does not exist
-        ('filedoesnotexist.tcst', []),
-        # file is empty
-        ('empty_filter.tcst', []),
-        # file has STORM_ID column with 4 values
-        ('fake_filter_20141214_00.tcst', ['ML1201072014',
-                                          'ML1221072014',
-                                          'ML1241072014',
-                                          'ML1251072014']),
-        # file does not have STORM_ID column
-        ('test_20190101.stat', []),
-    ]
-)
-@pytest.mark.util
-def test_get_storm_ids(metplus_config, filename, expected_result):
-    config = metplus_config
-    filepath = os.path.join(config.getdir('METPLUS_BASE'),
-                            'internal', 'tests',
-                            'data',
-                            'stat_data',
-                            filename)
-
-    assert util.get_storms(filepath, id_only=True) == expected_result
-
-
-@pytest.mark.parametrize(
-    'filename, expected_result', [
-        # file does not exist
-        ('filedoesnotexist.tcst', []),
-        # file is empty
-        ('empty_filter.tcst', []),
-        # file has STORM_ID column with 4 values
-        ('fake_filter_20141214_00.tcst', ['header',
-                                          'ML1201072014',
-                                          'ML1221072014',
-                                          'ML1241072014',
-                                          'ML1251072014']),
-        # file does not have STORM_ID column
-        ('test_20190101.stat', []),
-    ]
-)
-@pytest.mark.util
-def test_get_storms(metplus_config, filename, expected_result):
-    storm_id_index = 4
-    config = metplus_config
-    filepath = os.path.join(config.getdir('METPLUS_BASE'),
-                            'internal', 'tests',
-                            'data',
-                            'stat_data',
-                            filename)
-
-    storm_dict = util.get_storms(filepath)
-    print(storm_dict)
-    assert list(storm_dict.keys()) == expected_result
-    for storm_id in expected_result[1:]:
-        for storm_line in storm_dict[storm_id]:
-            # ensure storm_id_index matches storm ID
-            assert storm_line.split()[storm_id_index] == storm_id
-
-    # ensure header matches expected format
-    if storm_dict:
-        assert storm_dict['header'].split()[storm_id_index] == 'STORM_ID'
-
-
-@pytest.mark.util
-def test_get_storms_mtd(metplus_config):
-    index = 23
-    expected_result = [
-        'header',
-        'CF001',
-        'CO001'
-    ]
-    sort_column = 'OBJECT_CAT'
-    config = metplus_config
-    filepath = os.path.join(config.getdir('METPLUS_BASE'),
-                            'internal', 'tests',
-                            'data',
-                            'mtd',
-                            'fake_mtd_2d.txt')
-
-    storm_dict = util.get_storms(filepath, sort_column=sort_column)
-    print(storm_dict)
-    assert list(storm_dict.keys()) == expected_result
-    for storm_id in expected_result[1:]:
-        for storm_line in storm_dict[storm_id]:
-            # ensure index matches storm ID
-            assert storm_line.split()[index] == storm_id
-
-    # ensure header matches expected format
-    if storm_dict:
-        assert storm_dict['header'].split()[index] == sort_column
 
 
 @pytest.mark.parametrize(
